Issue 155_5: Merged deprecated JPAWebloggerImpl into Weblogger
