+++ app/src/main/resources/spring-beans.xml
@@ -23,7 +23,7 @@
     <!-- Have JSR-250 annotations (PostConstruct, PreDestroy) honored by Spring -->
     <beans:bean class="org.springframework.context.annotation.CommonAnnotationBeanPostProcessor"/>
 
-    <beans:bean id="webloggerBean" class="org.apache.roller.weblogger.business.jpa.JPAWebloggerImpl">
+    <beans:bean id="webloggerBean" class="org.apache.roller.weblogger.business.Weblogger">
         <beans:constructor-arg ref="indexManager"/>
         <beans:constructor-arg ref="mediaFileManager"/>
         <beans:constructor-arg ref="fileContentManager"/>
@@ -31,7 +31,6 @@
         <beans:constructor-arg ref="pluginManager"/>
         <beans:constructor-arg ref="propertiesManager"/>
         <beans:constructor-arg ref="themeManager"/>
-        <beans:constructor-arg ref="threadManager"/>
         <beans:constructor-arg ref="userManager"/>
         <beans:constructor-arg ref="weblogManager"/>
         <beans:constructor-arg ref="weblogEntryManager"/>
